Don't store the minhash implementation version in the db
The version depends on whatever version of the minhash implementation
(MHI) is available on the command line, which is not under the control
of the service. Also, since the sketch databases are provided by a user
and there's no way to know what version of the MHI was used to create
the database, assigning whatever version of the MHI happens to be
available on the command line to the namespace doesn't make any sense.