Only keep <max return count> distances in memory
Preparation for running mash against multiple files.

Adds a capped sorted collection that only performs a sort when a) the
collection size is below the cap or b) the item to be added to the
collection is larger than the smallest item in the collection (or vice
versa if descending is specified).

This is used to only keep the top <max return count> distance
measurements in memory at any one time, and means the sort is now n log
<max return count> rather than n log n. In practice many of the distance
measurements will be rejected immediately and no sort will occur.